+++ SanteDB.Core.Model/VersionedEntityData.cs
@@ -89,7 +89,7 @@ namespace SanteDB.Core.Model
         /// The sequence number of the version (for ordering)
         /// </summary>
         [XmlElement("sequence"), JsonProperty("sequence")]
-        public Int32? VersionSequence { get; set; }
+        public Int64? VersionSequence { get; set; }
 
         /// <summary>
         /// Represent the versioned data as a string
